refactor: global, myProjectHistory, myProjects contents 분리
